@@ -6,39 +6,45 @@ publish_to: 'none'
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.11.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  cupertino_icons: ^0.1.3
-  flutter_modular: ^2.0.1
+
+  flutter_modular: ^3.0.0-nullsafety.41
+  flutter_bloc: ^7.0.0-nullsafety.1
+  bloc: ^7.0.0-nullsafety.1
+  rxdart: ^0.26.0-nullsafety.1
+
   dio: ^3.0.10
-  rxdart: ^0.24.1
-  font_awesome_flutter: ^8.8.1
+  sembast: ^3.0.0-nullsafety.4
+
+  uuid: ^3.0.0-nullsafety.0
+  flutter_svg: 0.20.0-nullsafety.3
+
+  google_fonts: ^1.1.2
+  font_awesome_flutter: ^9.0.0-nullsafety
+  cupertino_icons: ^0.1.3
+
   qr_flutter: ^3.2.0
-  shared_preferences: ^0.5.12
-  google_fonts: ^1.1.0
+  qr_code_scanner: ^0.3.2
+
+  path: ^1.8.0-nullsafety.3
+  path_provider: ^2.0.0-nullsafety
   flutter_secure_storage: ^3.3.5
-  qr_code_scanner: ^0.0.13
-  flutter_svg: 0.19.0
+
   didkit:
     path: ../didkit/lib/flutter
-  bloc: ^6.1.0
-  flutter_bloc: ^6.1.1
-  sembast: ^2.4.8
-  path: ^1.7.0
-  path_provider: ^1.6.24
-  uuid: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   #intl_translation: ^0.17.10
-  pedantic: ^1.9.0
-  flutter_launcher_icons: ^0.8.1
+  pedantic: ^1.10.0
+  #flutter_launcher_icons: ^0.8.1
 
 flutter_icons:
   android: true
