@@ -22,7 +22,7 @@ class CredentialsList extends StatelessWidget {
     final localizations = AppLocalizations.of(context);
 
     return BlocListener(
-      cubit: Modular.get<ScanBloc>(),
+      bloc: Modular.get<ScanBloc>(),
       listener: (context, state) {
         if (state is ScanStateMessage) {
           ScaffoldMessenger.of(context).showSnackBar(SnackBar(
