@@ -1,6 +1,9 @@
+import 'package:credible/app/app_module.dart';
+import 'package:credible/app/app_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_modular/flutter_modular.dart';
 
-import 'app/app_module.dart';
-
-void main() => runApp(ModularApp(module: AppModule()));
+void main() => runApp(ModularApp(
+      module: AppModule(),
+      child: AppWidget(),
+    ));
